python-ai/requirements.txt
openai-whisper
torch<2.5
transformers
sentencepiece
